// src/modules/user/types/user.type.ts
export enum UserType {
  REVIEWER = 'reviewer',
  AUTHOR = 'author',
  SUPERADMIN = "super-admin",
  EDITOR_IN_CHIEF = "Editor-in-Chief",
  MANAGING_EDITOR = "Managing-Editor",
  SECTION_EDITOR = "Section-Editor",
  ASSOCIATE_EDITOR = "Associate-Editor",
  PRODUCTION_EDITOR = "Production-Editor",
  COPY_EDITOR = "Copy-Editor"
}

export enum EditorRole {
  EDITOR_IN_CHIEF = "Editor-in-Chief",
  MANAGING_EDITOR = "Managing-Editor",
  SECTION_EDITOR = "Section-Editor",
  ASSOCIATE_EDITOR = "Associate-Editor",
  GUEST_EDITOR = "Guest-Editor",
  COPY_EDITOR = "Copy-Editor"
}
export enum Sections{
  SCIENCES,
  COMPUTING,
  AGRICULTURE,
  CHEMISTRY
}